requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
docstring_parser==0.16
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
importlib_metadata==7.1.0
instructor==1.2.4
Jinja2==3.1.3
joblib==1.4.0
litellm==1.35.33
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
openai==1.25.0
packaging==24.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
shellingham==1.5.4
sniffio==1.3.1
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.19.1
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
zipp==3.18.1
